requirements.txt
fastapi
uvicorn
supabase
pandas
scikit-learn
python-dotenv
requests
numpy
joblib
